fix: split asset processing module
