+++ tools/asset-processor/Cargo.toml
@@ -5,10 +5,7 @@ edition = "2024"
 
 [dependencies]
 clap = { workspace = true, features = ["derive"] }
-peridot-rendering-configuration = { path = "../../modules/rendering-configuration", features = ["compilation"] }
-image = "0.25.8"
-ktx = { path = "../../thirdparty/ktx" }
-bedrock = { workspace = true }
+peridot-asset-processing = { path = "../../modules/asset-processing" }
+peridot-rendering-configuration = { path = "../../modules/rendering-configuration", features = ["with-asset-processing"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["fmt", "env-filter"] }
-thiserror = { workspace = true }
